+++ mxops/execution/steps.py
@@ -6,6 +6,7 @@ This module contains the classes used to execute scenes in a scenario
 from __future__ import annotations
 import base64
 from dataclasses import dataclass, field
+from importlib.util import spec_from_file_location, module_from_spec
 import os
 from pathlib import Path
 import sys
@@ -367,10 +368,11 @@ class ContractQueryStep(Step):
     arguments: List = field(default_factory=lambda: [])
     expected_results: List[Dict[str, str]] = field(default_factory=lambda: [])
     print_results: bool = False
+    results: List[QueryResult] | None = field(init=False, default=None)
 
-    def _interpret_return_data(self, data: str) -> str | QueryResult:
+    def _interpret_return_data(self, data: str) -> QueryResult:
         if not data:
-            return data
+            return QueryResult('', '', None)
 
         try:
             as_bytes = base64.b64decode(data)
@@ -400,29 +402,29 @@ class ContractQueryStep(Step):
         query = builder.build()
         proxy = ProxyNetworkProvider(config.get("PROXY"))
 
-        results = []
         results_empty = True
         n_attempts = 0
         max_attempts = int(Config.get_config().get("MAX_QUERY_ATTEMPTS"))
         while results_empty and n_attempts < max_attempts:
             n_attempts += 1
             response = proxy.query_contract(query)
-            results = [self._interpret_return_data(rsp) for rsp in response.return_data]
-            results_empty = (len(results) == 0 or
-                             (len(results) == 1 and results[0] == ""))
+            self.results = [
+                self._interpret_return_data(data) for data in response.return_data
+            ]
+            results_empty = (len(self.results) == 0 or
+                             (len(self.results) == 1 and self.results[0] == ""))
             if results_empty:
                 time.sleep(3)
                 LOGGER.warning(
                     f'Empty query result, retrying. Attempt {n_attempts}/{max_attempts}'
                 )
-
         if results_empty:
             raise errors.EmptyQueryResults
         if self.print_results:
-            print(results)
+            print(self.results)
         if len(self.expected_results) > 0:
             LOGGER.info("Saving Query results as contract data")
-            for result, expected_result in zip(results, self.expected_results):
+            for result, expected_result in zip(self.results, self.expected_results):
                 parsed_result = parse_query_result(
                     result, expected_result["result_type"]
                 )
@@ -1070,3 +1072,51 @@ def instanciate_steps(raw_steps: List[Dict]) -> List[Step]:
             raise errors.InvalidStepDefinition(step_type, raw_step) from err
         steps_list.append(step)
     return steps_list
+
+
+@dataclass
+class PythonStep(Step):
+    """
+    This Step execute a custom python function of the user
+    """
+    module_path: str
+    function: str
+    arguments: list = field(default_factory=list)
+    keyword_arguments: dict = field(default_factory=dict)
+
+    def execute(self):
+        """
+        Execute the specified function
+        """
+        module_path = Path(self.module_path)
+        module_name = module_path.stem
+        LOGGER.info(
+            f"Executing python function {self.function} from user module {module_name}"
+            )
+
+        # load module and function
+        spec = spec_from_file_location(
+            module_name,
+            module_path.as_posix()
+        )
+        user_module = module_from_spec(spec)
+        spec.loader.exec_module(user_module)
+        user_function = getattr(user_module, self.function)
+
+        # transform args and kwargs and execute
+        arguments = [utils.retrieve_value_from_any(arg) for arg in self.arguments]
+        keyword_arguments = {
+            key: utils.retrieve_value_from_any(val)
+            for key, val in self.keyword_arguments.items()
+        }
+        result = user_function(*arguments, **keyword_arguments)
+
+        if result:
+            if isinstance(result, str):
+                var_name = f"MXOPS_{self.function.upper()}_RESULT"
+                os.environ[var_name] = result
+            else:
+                LOGGER.warning(f"The result of the function {self.function} is not a "
+                               "string and has not been saved")
+
+        LOGGER.info(f"Function result: {result}")